refactor: jsonDeserializer 위치 변경
